Fixed clang-analyzer-cplusplus and Visual Studio c++ Linter issue.

+++ src/Tuner.cpp
@@ -257,7 +257,7 @@ std::string Tuner<net_t>::parameters_to_string(const Parameters& p) {
 }
 
 static size_t next_power_of_two(const size_t x) {
-    return 2 << size_t(std::ceil(std::log2(x)) - 1);
+    return static_cast<size_t>(2) << size_t(std::ceil(std::log2(x)) - 1);
 }
 
 template <typename net_t>
